# public/api-docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Task Manager API",
        "description": "REST API for task management with role-based access control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/proj1/public/index.php?route=",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/tasks/all": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "operationId": "40b7fe0ed29f4054be573cdc59c400ae",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task (admin only)",
                "operationId": "1282b45220f504e39953e09f038bf992",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task": {
                                        "type": "string",
                                        "example": "Complete the project"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Developer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Task and role required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (admin access required)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "operationId": "83d3a8edddbcdc82d4a663d90513e49f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "string"
                                    },
                                    "employee_responsible": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "task_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks/{task_id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task (admin only)",
                "operationId": "7fbe6f772dca7e4fcb003065fbc82a01",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Task ID required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (admin access required)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks/{task_id}/take": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Take a task",
                "operationId": "9bb69ff054e38d2551962a0d3a93b046",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "ID of the task to take",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task taken",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Task ID required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks/{task_id}/complete": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Complete a task",
                "operationId": "b6576f2a9a4fc736195ac8bceee718b4",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "ID of the task to complete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Task ID required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks/my": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks assigned to the current user",
                "operationId": "6591582072524b9534d4212e72ba11cb",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/tasks/team": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks assigned to the current user's team",
                "operationId": "081899081b8bab0b00179959a9d5df3b",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the task",
                        "type": "integer"
                    },
                    "task": {
                        "description": "Description of the task",
                        "type": "string"
                    },
                    "assigned_to": {
                        "description": "Team assigned to the task",
                        "type": "string"
                    },
                    "employee_responsible": {
                        "description": "Username of the employee responsible for the task",
                        "type": "string"
                    },
                    "status": {
                        "description": "Current status of the task (Pending, In Progress, Completed)",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the task was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the task was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "description": "Session-based authentication via PHP session cookie",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}